Update version 1.1.9 and fix issue with too long integers as subject identifier
app_server.R:
-fix font size of subject identifier on y-axis for main plot and zoom-panel
-add new observer to react when new button "reset panel position" is clicked (re-create panels)

app_ui.R:
- add actionButton 'reset_draggable_panel_positions'

updates_variables_selection_csvA.R:
-change 'is.integer' to is.numeric to avoid stricter restriction

updates_variables_selection_csvB.R:
-change 'is.integer' to is.numeric to avoid stricter restriction

updates_variables_selection_file.R:
-change 'is.integer' to is.numeric to avoid stricter restriction

updates_variables_selection_rdataA.R:
-change 'is.integer' to is.numeric to avoid stricter restriction

updates_variables_selection_rdataB.R:
-change 'is.integer' to is.numeric to avoid stricter restriction

utils_numeric_to_integer.R:
-change 'is.integer' to is.numeric to avoid stricter restriction
-remove restriction that subject identifier can not be character

Change test files to match new updates

+++ R/updates_variables_selection_rdataA.R
@@ -17,7 +17,7 @@ updates_variables_selection_rdataA <- function(
           A <- get(load(file$datapath))
           A2 <- numeric_to_integer(A)
 
-          integers_A <- names(which(unlist(lapply(A2,is.integer))))
+          integers_A <- names(which(unlist(lapply(A2,is.numeric))))
 
           if (shiny::isRunning()) {
             shiny::updateSelectInput(
